paspa

+++ src/main/java/com/example/demo/common/tiktok/model/order/TiktokOrderModel.java
@@ -87,5 +87,5 @@ public class TiktokOrderModel implements Serializable {
     private long updateTime;
 
     @JsonProperty("paid_time")
-    private int paidTime;
+    private long paidTime;
 }
